requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.7.0
pydantic==2.5.0
```

**File 2: `Procfile`** (no extension)
```
web: uvicorn main:app --host 0.0.0.0 --port $PORT
```

**File 3: `.gitignore`**
```
__pycache__/
*.pyc
.env
venv/
*.db
cache/
conversation_memory/